Tuned *install.sh scripts. Minor edits to aliases/arch install scripts.

+++ uninstall.sh
@@ -1,46 +1,35 @@
 #! /bin/bash
 
-### Uninstall Pip3 packages
-#pip3 uninstall		\
-#  thefuck
-
-### Uninstall APT packages
-#sudo sh -c "
-#  apt --purge remove	\
-#    facter		\
-#    libnotify-bin	\
-#    ntp			\
-#    python3		\
-#    python3-dev		\
-#    python3-pip		\
-#    python3-venv	\
-#    safe-rm		\
-#    sl			\
-#    snapd		\
-#    speedtest-cli	\
-#    thefuck		\
-#    tldr		\
-#    lm-sensors -y;
-#
-#  apt --purge autoremove -y;
-#  apt autoclean;
-#"
-
 ### Delete installed profile if it exists and replace with saved profile.
 if [ $(head -n 1 ~/.profile)="###transnat" ]; then
-  rm ~/.profile
-  mv ~/.profile.original ~/.profile
+	rm ~/.profile
+  
+	if [ -f ~/.profile.original ]; then
+		mv ~/.profile.original ~/.profile
+	fi
 fi
 
 ### Delete installed bashrc if it exists and replace with saved bashrc.
 if [ $(head -n 1 ~/.bashrc)="###transnat" ]; then
-  rm ~/.bashrc
-  mv ~/.bashrc.original ~/.bashrc
+	rm ~/.bashrc
+  
+	if [ -f ~/.bashrc.original ]; then
+		mv ~/.bashrc.original ~/.bashrc
+	fi
+fi
+
+### Delete installed screenrc if it exists and replace with saved screenrc.
+if [ $(head -n 1 ~/.screenrc)="###transnat" ]; then
+	rm ~/.screenrc
+  
+	if [ -f ~/.screenrc.original ]; then
+		mv ~/.screenrc.original ~/.screenrc
+	fi
 fi
 
 ### Delete profile folder if it exists.
-if [ -d ~/.bash_profile.d ]; then
-  rm -rf ~/.bash_profile.d
+if [ -d ~/.bash ]; then
+	rm -rf ~/.bash
 fi
 
 ### Load old profile.
